Modify 'nasm--opt' for supporting parametric keywords
After this commit, parametric keywords like "{foo}", "{foo=}" and
"{foo=bar}" are highlighted, matching NASM's behavior.

Add 'nasm--parametric-keyword-regexp' constant and 'nasm--split',
'nasm--is-parametric-keyword' and 'nasm--get-parametric-regexp'
functions. These are defined in an 'eval-when-compile' block, since they
are only used by 'nasm--opt'; this is not ideal.

Modify NASM syntax table to include "{" and "}" as symbols, since they
are part of regexps matched through "\\_< ... \\_>". Although these
characters are not normally used when programming in NASM, this might
have unintended consequences, since it essentially removes two symbol
delimiters.